+++ lib/ironman/config.ex
@@ -4,67 +4,70 @@ defmodule Ironman.Config do
   all the checks, where it is updated, and then files are written out at the end based on its contents.
   """
 
-  alias Ironman.Config
+  alias Ironman.{Config, Utils}
   alias Ironman.Utils.File, as: IFile
+
+  @settable_keys [
+    :mix_exs,
+    :gitignore,
+    :dialyzer_ignore,
+    :config_exs,
+    :config_test_exs,
+    :config_dev_exs,
+    :config_prod_exs
+  ]
 
   @type t :: %__MODULE__{
           mix_exs: String.t(),
-          mix_exs_changed: boolean(),
           gitignore: String.t() | nil,
-          gitignore_changed: boolean(),
           dialyzer_ignore: String.t() | nil,
-          dialyzer_ignore_changed: boolean(),
-          starting_project_config: keyword()
+          config_exs: String.t(),
+          config_test_exs: String.t() | nil,
+          config_dev_exs: String.t() | nil,
+          config_prod_exs: String.t() | nil,
+          starting_project_config: keyword(),
+          changed: MapSet.t(atom())
         }
   defstruct mix_exs: nil,
-            mix_exs_changed: false,
             gitignore: nil,
-            gitignore_changed: false,
             dialyzer_ignore: nil,
-            dialyzer_ignore_changed: false,
-            starting_project_config: nil
+            config_exs: nil,
+            config_test_exs: nil,
+            config_dev_exs: nil,
+            config_prod_exs: nil,
+            starting_project_config: nil,
+            changed: MapSet.new()
 
   def new!() do
     %__MODULE__{
-      mix_exs: file_or_nil("mix.exs"),
-      gitignore: file_or_nil(".gitignore"),
-      dialyzer_ignore: file_or_nil(".dialyzer_ignore.exs"),
+      mix_exs: file_or_nil(Utils.path_of(:mix_exs)),
+      config_exs: file_or_nil(Utils.path_of(:config_exs)),
+      config_dev_exs: file_or_nil(Utils.path_of(:config_dev_exs)),
+      config_test_exs: file_or_nil(Utils.path_of(:config_test_exs)),
+      config_prod_exs: file_or_nil(Utils.path_of(:config_prod_exs)),
+      gitignore: file_or_nil(Utils.path_of(:gitignore)),
+      dialyzer_ignore: file_or_nil(Utils.path_of(:dialyzer_ignore)),
       starting_project_config: Mix.Project.config()
     }
   end
 
-  ## Getters
-  def starting_project_config(%Config{starting_project_config: starting_project_config}), do: starting_project_config
-  def mix_exs(%Config{mix_exs: mix_exs}), do: mix_exs
-  def gitignore(%Config{gitignore: gitignore}), do: gitignore
-  def dialyzer_ignore(%Config{dialyzer_ignore: dialyzer_ignore}), do: dialyzer_ignore
+  def get(%Config{} = config, key), do: Map.get(config, key)
 
-  ## Setters when not actually changed
+  @spec set(Config.t(), atom(), String.t(), boolean()) :: Config.t()
+  def set(%Config{changed: changed} = config, key, value, changed_flag \\ true) when key in @settable_keys do
+    config =
+      if changed_flag and get(config, key) != value do
+        %Config{config | changed: MapSet.put(changed, key)}
+      else
+        config
+      end
 
-  def set_mix_exs(%Config{mix_exs: mix_exs} = config, new_content)
-      when new_content == mix_exs,
-      do: config
+    Map.put(config, key, value)
+  end
 
-  def set_mix_exs(%Config{} = config, new_content),
-    do: %Config{config | mix_exs: new_content, mix_exs_changed: true}
+  def changed?(%Config{changed: changed}, key), do: MapSet.member?(changed, key)
 
-  def set_gitignore(%Config{gitignore: gitignore} = config, new_content)
-      when new_content == gitignore,
-      do: config
-
-  def set_gitignore(%Config{} = config, new_content),
-    do: %Config{config | gitignore: new_content, gitignore_changed: true}
-
-  def set_dialyzer_ignore(%Config{dialyzer_ignore: dialyzer_ignore} = config, new_content)
-      when new_content == dialyzer_ignore,
-      do: config
-
-  def set_dialyzer_ignore(%Config{} = config, new_content),
-    do: %Config{config | dialyzer_ignore: new_content, dialyzer_ignore_changed: true}
-
-  def mix_exs_changed(%Config{mix_exs_changed: mix_exs_changed}), do: mix_exs_changed
-  def gitignore_changed(%Config{gitignore_changed: gitignore_changed}), do: gitignore_changed
-  def dialyzer_ignore_changed(%Config{dialyzer_ignore_changed: dialyzer_ignore_changed}), do: dialyzer_ignore_changed
+  def any_changed?(%Config{changed: changed}), do: MapSet.size(changed) > 0
 
   def app_name(%Config{starting_project_config: starting_project_config}),
     do: Keyword.fetch!(starting_project_config, :app)