+++ lib/ironman/utils/cmd/default_impl.ex
@@ -2,9 +2,11 @@ defmodule Ironman.Utils.Cmd.DefaultImpl do
   @moduledoc false
   @behaviour Ironman.Utils.Cmd.Impl
 
-  @spec run([String.t()]) :: :ok
+  @spec run([String.t()]) :: :ok | :error
   def run([h | t]) do
-    {_, 0} = System.cmd(h, t)
-    :ok
+    case System.cmd(h, t) do
+      {_, 0} -> :ok
+      _ -> :error
+    end
   end
 end