+++ test/support/mox_helpers.ex
@@ -16,14 +16,9 @@ defmodule Ironman.Test.Helpers.MoxHelpers do
     |> expect(:get, fn ^expect -> "#{return}\n" end)
   end
 
-  def raise_on_io do
-    Ironman.MockIO
-    |> expect(:get, fn x -> raise "IO.get(\"#{x}\")" end)
-  end
-
-  def expect_cmd(expect) do
+  def expect_cmd(expect, return \\ :ok) do
     Ironman.MockCmd
-    |> expect(:run, fn ^expect -> :ok end)
+    |> expect(:run, fn ^expect -> return end)
   end
 
   def expect_file_exists?(file, exists? \\ true) do
@@ -31,35 +26,13 @@ defmodule Ironman.Test.Helpers.MoxHelpers do
     |> expect(:exists?, fn ^file -> exists? end)
   end
 
-  def raise_on_file_exists? do
-    Ironman.MockFile
-    |> expect(:exists?, fn x -> raise "File.exists?(\"#{x}\")" end)
-  end
-
   def expect_file_read!(file, content) do
     Ironman.MockFile
     |> expect(:read!, fn ^file -> content end)
   end
 
-  def raise_on_file_read! do
-    Ironman.MockFile
-    |> expect(:read!, fn x -> raise "File.read!(\"#{x}\")" end)
-  end
-
   def expect_file_write!(file, contents) do
     Ironman.MockFile
     |> expect(:write!, fn ^file, ^contents -> :ok end)
-  end
-
-  def raise_on_file_write! do
-    Ironman.MockFile
-    |> expect(:write!, fn x, y -> raise "File.write!(\"#{x}\", \"#{y}\")" end)
-  end
-
-  def raise_on_any_other do
-    raise_on_file_exists?()
-    raise_on_file_read!()
-    raise_on_file_write!()
-    raise_on_io()
   end
 end